support levelDB storage
only support paranoid option

+++ ruby/server/lib/roma/storage/leveldb_storage.rb
@@ -6,33 +6,33 @@ module Roma
 
     class LevelDBStorage < BasicStorage
       include LevelDB
-      
+
       class LevelDB::DB
         alias get_org get
         def get key
           begin
             ret = get_org key
-          rescue => ecode
+          rescue Exception => ecode
             raise StorageException, errmsg(ecode)
           end
           ret
         end
-        
+
         alias put_org put
         def put key, value
           begin
             ret = put_org key, value
-          rescue => ecode
+          rescue Exception => ecode
             raise StorageException, errmsg(ecode) unless ret
           end
           ret
         end
-        
+
         alias out_org delete
         def out key
           begin
             ret = out_org key
-          rescue => ecode
+          rescue Exception => ecode
             raise StorageException, errmsg(ecode)
           end
           ret
@@ -42,7 +42,7 @@ module Roma
         def rnum
           begin
             ret = rnum_org
-          rescue => ecode
+          rescue Exception => ecode
             raise StorageException, errmsg(ecode)
           end
           ret
@@ -91,71 +91,46 @@ module Roma
         buf = @hdb; @hdb = []
         buf.each{ |hdb| close_db(hdb) }
 
-        File.unlink(@fname_lock) if @fname_lock
-        @fname_lock = nil
-      end
-
-      protected
-
-      def set_options(hdb)
-        #prop = parse_options
-
-        #prop.each_key{|k|
-        #  unless /^(bnum|apow|fpow|opts|xmsiz|rcnum|dfunit)$/ =~ k
-        #    raise RuntimeError.new("Syntax error, unexpected option #{k}")
-        #  end
-        #}
-        
-        #opts = 0
-        #if prop.key?('opts')
-        #  opts |= HDB::TLARGE if prop['opts'].include?('l')
-        #  opts |= HDB::TDEFLATE if prop['opts'].include?('d')
-        #  opts |= HDB::TBZIP if prop['opts'].include?('b')
-        #  opts |= HDB::TTCBS if prop['opts'].include?('t')
-        #end
-
-        #hdb.tune(prop['bnum'].to_i,prop['apow'].to_i,prop['fpow'].to_i,opts)
-
-        #hdb.setxmsiz(prop['xmsiz'].to_i) if prop.key?('xmsiz')
-        #hdb.setcache(prop['rcnum'].to_i) if prop.key?('rcnum')
-        #hdb.setdfunit(prop['dfunit'].to_i) if prop.key?('dfunit')
+        if @fname_lock
+          if File.exist?(@fname_lock)
+            File.unlink(@fname_lock) if @fname_lock
+          end
+          @fname_lock = nil
+        end
       end
 
       private
-
       def parse_options
-        #return Hash.new(-1) unless @option
-        #buf = @option.split('#')
-        #prop = Hash.new(-1)
-        #buf.each{|equ|
-        #  if /(\S+)\s*=\s*(\S+)/ =~ equ
-        #    prop[$1] = $2
-        #  else
-        #    raise RuntimeError.new("Option string parse error.")
-        #  end
-        #}
-        #prop
+        return {} unless @option
+
+        opt = {}
+        buf = @option.split('#')
+        buf.each do |equ|
+          if /(\S+)\s*=\s*(\S+)/ =~ equ
+            key = $1.to_sym
+            if $2 == "true"
+              opt[key] = true
+            elsif $2 == "false"
+              opt[key] = false
+            elsif $2 =~ /^[0-9]+%/
+              opt[key] = $2.to_i
+            else
+              opt[key] = $2
+            end
+          else
+            raise RuntimeError.new("Option string parse error.")
+          end
+        end
+
+        opt
       end
 
       def open_db(fname)
-        #begin
-        #LevelDB::DB::new fname
-        #rescue => ecode
-        #  raise RuntimeError.new("ldb open error #{ecode}")
-        #end
-
-        #set_options(hdb)
-        
-        #if !hdb.open(fname, HDB::OWRITER | HDB::OCREAT | HDB::ONOLCK)
-        #  ecode = hdb.ecode
-        #  raise RuntimeError.new("ldb open error")
-        #end
-
-        #hdb
         begin
-          LevelDB::DB::new fname
+          @parsed_option = parse_options
+          LevelDB::DB.new(fname, @parsed_option)
         rescue => ecode
-          raise RuntimeError.new("ldb open error")
+          raise RuntimeError, "ldb open error: #{ecode.to_s}", ecode.backtrace
         end
       end
